+++ src/main/java/com/app/eventmanager/repository/RewardRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.eventmanager.model.Reward;
 
-public interface RewardRepository extends JpaRepository<Reward, Integer>{
+public interface RewardRepository extends JpaRepository<Reward, Long>{
 
 
 }
